chore: restructure shared ui package
