+++ packages/ui/postcss.config.mjs
@@ -1,0 +1,7 @@
+const config = {
+  plugins: {
+    autoprefixer: {},
+  },
+};
+
+export default config;
