+++ packages/ui/src/components/collapsible.tsx
