+++ packages/ui/src/components/resizable.tsx
